refactor: Calendar 함수 utils화
